Bclerc

+++ back/src/chat/services/penalties/penalties.service.ts
@@ -28,7 +28,6 @@ export class PenaltiesService {
         endTime: end_at ? end_at : new Date(),
       }
     });
-    console.log(ret);
   }
 
   async getPenaltyById(id: number): Promise<ChatPenalty> {
@@ -48,6 +47,8 @@ export class PenaltiesService {
   }
 
   async deletePenalty(id: number) {
+    if (!id)
+      return ;
     await this.prisma.chatPenalty.delete({
       where: {
         id: id
@@ -56,7 +57,6 @@ export class PenaltiesService {
   }
 
   async getRoomPenaltiesForUser(userId: number, roomId: number): Promise<ChatPenalty> {
-    console.log(userId, roomId);
     const penalties = await this.prisma.chatPenalty.findFirst({
       where: {
         userId: userId,
